+++ src/components/LayoutMain/index.type.ts
@@ -1,0 +1,5 @@
+import { ReactNode } from "react";
+
+export type LayoutMainProps = {
+  children: ReactNode;
+};
