+++ src/components/Meta/index.type.ts
@@ -1,0 +1,4 @@
+export type MetaProps = {
+  title?: string;
+  description?: string;
+};
