+++ src/components/Pagination/index.tsx
@@ -1,71 +1,24 @@
 import Link from "next/link";
-
-type PaginationType = {
-  currentPage: number | any;
-  totalPages?: number;
-  path: string;
-};
+import { usePagination } from "./usePagination";
+import { useUrlSetting } from "./useUrlSetting";
+import { PaginationProps } from "./index.type";
 
 export const Pagination = ({
   currentPage,
   path,
   totalPages,
-}: PaginationType) => {
-  let rootUrl;
-  if (process.env.NODE_ENV === "production") {
-    //本番
-    rootUrl = process.env.NEXT_PUBLIC_SITE_URL;
-  } else if (process.env.NODE_ENV === "development") {
-    //ローカル
-    rootUrl = "http://localhost:3000";
-  }
+}: PaginationProps) => {
+  const { resultPath, pageParam, prevPage, nextPage, resultTotalPage } =
+    useUrlSetting({ currentPage, path, totalPages });
 
-  //pageパラメータを除く
-  let url = new URL(path, rootUrl);
-  url.searchParams.delete("page");
-  const resultPath = url.href;
-
-  //pageパラメータの書き方
-  let pageParam = "&page=";
-  if (path === "/") {
-    //トップページの場合
-    pageParam = "?page=";
-  }
-
-  if (!totalPages) totalPages = 1;
-
-  const prevPage = Number(currentPage) - 1;
-  const nextPage = Number(currentPage) + 1;
+  const { pageNumbers } = usePagination({
+    currentPage,
+    totalPages: resultTotalPage,
+  });
 
   const liStyle =
     "flex justify-center items-center w-[40px] h-[40px] border text-center text-[14px] border-[#F8E6CC]";
   const currentStyle = "bg-[#017D01] text-white pointer-events-none";
-
-  /**
-   * 動的な数字部分
-   */
-  const renderPageNumbers = () => {
-    const pageNumbers = [];
-    const startPage = Math.max(1, currentPage - 1);
-    const endPage = Math.min(totalPages!, startPage + 2);
-
-    for (let i = startPage; i <= endPage; i++) {
-      pageNumbers.push(
-        <li key={i}>
-          <Link
-            href={`${resultPath}${pageParam}${i}`}
-            className={`dynamicPage rounded ${liStyle} ${
-              i === Number(currentPage) ? currentStyle : ""
-            }`}
-          >
-            {i}
-          </Link>
-        </li>
-      );
-    }
-
-    return pageNumbers;
-  };
 
   return (
     <section>
@@ -100,20 +53,31 @@ export const Pagination = ({
         )}
 
         {/* 動的な数字部分 */}
-        {renderPageNumbers()}
+        {pageNumbers.map((page) => (
+          <li key={page.key}>
+            <Link
+              href={`${resultPath}${pageParam}${page.key}`}
+              className={`dynamicPage rounded ${liStyle} ${
+                page.key === Number(currentPage) ? currentStyle : ""
+              }`}
+            >
+              {page.key}
+            </Link>
+          </li>
+        ))}
 
         {/* 最後 */}
-        {nextPage <= totalPages - 1 && (
+        {nextPage <= resultTotalPage - 1 && (
           <>
             <li>
               <span className={`${liStyle} rounded`}>...</span>
             </li>
             <li>
               <Link
-                href={`${resultPath}${pageParam}${totalPages}`}
+                href={`${resultPath}${pageParam}${resultTotalPage}`}
                 className={`${liStyle} rounded`}
               >
-                {totalPages}
+                {resultTotalPage}
               </Link>
             </li>
           </>
@@ -121,7 +85,7 @@ export const Pagination = ({
 
         {/* 次へ */}
         <li>
-          {nextPage <= totalPages && (
+          {nextPage <= resultTotalPage && (
             <Link
               href={`${resultPath}${pageParam}${nextPage}`}
               className={`${liStyle} rounded`}
