+++ src/components/Pagination/index.type.ts
@@ -1,0 +1,16 @@
+export type PaginationProps = {
+  currentPage: number | any;
+  totalPages: number;
+  path: string;
+};
+
+export type UsePaginationProps = {
+  currentPage: number;
+  totalPages: number;
+};
+
+export type UseUrlSettingProps = {
+  currentPage: number;
+  path: string;
+  totalPages: number;
+};
