+++ src/components/SearchArea/index.type.ts
@@ -1,0 +1,4 @@
+export type SearchAreaProps = {
+  prefecture: [];
+  genres: [];
+};
