+++ src/components/TextArea/index.type.ts
@@ -1,0 +1,6 @@
+export type TextAreaProps = {
+  isLoad?: boolean;
+  searchNum: number | null;
+  totalPages: number;
+  currentPage: any;
+};
