+++ src/components/TopMv/index.type.ts
@@ -1,0 +1,4 @@
+export type TopMvProps = {
+  prefecture: [];
+  genres: [];
+};
